+++ packages/cmc/src/cmc.options.ts
@@ -4,5 +4,5 @@ import type { _SERVICE as CanisterService } from "../candid/cmc";
 
 export interface CMCCanisterOptions
   extends Omit<CanisterOptions<CanisterService>, "canisterId"> {
-  canisterId: string | Principal;
+  canisterId: Principal;
 }